Refactored code generation to its own interface
