+++ Source/Editor/AGS.Editor/CodeGen/ICodeGenerator.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Text;
+
+namespace AGS.Editor
+{
+    public interface ICodeGenerator
+    {
+        void GenerateCode(EntityModel model, StringBuilder code);
+    }
+}
